codeaudit/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-sql-connector==1.4.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
langchain-google-vertexai==0.0.6
pylint==3.0.2
bandit==1.7.5
radon==6.0.1